# backend/openapi.json
{
  "components": {
    "schemas": {
      "Body_login_api_v1_auth_login_post": {
        "properties": {
          "client_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Id"
          },
          "client_secret": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "format": "password",
            "title": "Client Secret"
          },
          "grant_type": {
            "anyOf": [
              {
                "pattern": "^password$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grant Type"
          },
          "password": {
            "format": "password",
            "title": "Password",
            "type": "string"
          },
          "scope": {
            "default": "",
            "title": "Scope",
            "type": "string"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ],
        "title": "Body_login_api_v1_auth_login_post",
        "type": "object"
      },
      "EnumOption": {
        "description": "A selectable option. `label` is display text derived from the value.",
        "properties": {
          "label": {
            "title": "Label",
            "type": "string"
          },
          "value": {
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "value",
          "label"
        ],
        "title": "EnumOption",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "LoginRequest": {
        "description": "JSON login, alongside the OAuth2 form endpoint.",
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "password": {
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
      "Token": {
        "description": "OAuth2 bearer token response.",
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "expires_in_minutes": {
            "title": "Expires In Minutes",
            "type": "integer"
          },
          "token_type": {
            "default": "bearer",
            "title": "Token Type",
            "type": "string"
          }
        },
        "required": [
          "access_token",
          "expires_in_minutes"
        ],
        "title": "Token",
        "type": "object"
      },
      "UserCreate": {
        "description": "Admin-only user creation.",
        "properties": {
          "department": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Department"
          },
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "full_name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Full Name",
            "type": "string"
          },
          "password": {
            "maxLength": 72,
            "minLength": 8,
            "title": "Password",
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole",
            "default": "viewer"
          }
        },
        "required": [
          "email",
          "full_name",
          "password"
        ],
        "title": "UserCreate",
        "type": "object"
      },
      "UserRead": {
        "description": "A user as returned by the API. Never includes the password hash.",
        "properties": {
          "department": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Department"
          },
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "full_name": {
            "title": "Full Name",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          }
        },
        "required": [
          "id",
          "email",
          "full_name",
          "role",
          "is_active"
        ],
        "title": "UserRead",
        "type": "object"
      },
      "UserRole": {
        "description": "Access roles. Ordering here is not privilege ordering - see core.rbac.",
        "enum": [
          "admin",
          "qa_manager",
          "complaint_officer",
          "investigator",
          "viewer"
        ],
        "title": "UserRole",
        "type": "string"
      },
      "UserWithPermissions": {
        "description": "`/auth/me` response.\n\nPermissions are sent so the UI can hide controls the user cannot use. This is\npresentation only - every endpoint enforces independently, and a tampered\nclient gains nothing.",
        "properties": {
          "department": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Department"
          },
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "full_name": {
            "title": "Full Name",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "permissions": {
            "items": {
              "type": "string"
            },
            "title": "Permissions",
            "type": "array"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          }
        },
        "required": [
          "id",
          "email",
          "full_name",
          "role",
          "is_active"
        ],
        "title": "UserWithPermissions",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VocabularyResponse": {
        "description": "Every enum the intake form and filters need, in one round trip.",
        "properties": {
          "complaint_sources": {
            "items": {
              "$ref": "#/components/schemas/EnumOption"
            },
            "title": "Complaint Sources",
            "type": "array"
          },
          "complaint_statuses": {
            "items": {
              "$ref": "#/components/schemas/EnumOption"
            },
            "title": "Complaint Statuses",
            "type": "array"
          },
          "complaint_types": {
            "items": {
              "$ref": "#/components/schemas/EnumOption"
            },
            "title": "Complaint Types",
            "type": "array"
          },
          "document_kinds": {
            "items": {
              "$ref": "#/components/schemas/EnumOption"
            },
            "title": "Document Kinds",
            "type": "array"
          },
          "dosage_forms": {
            "items": {
              "$ref": "#/components/schemas/EnumOption"
            },
            "title": "Dosage Forms",
            "type": "array"
          },
          "priorities": {
            "items": {
              "$ref": "#/components/schemas/EnumOption"
            },
            "title": "Priorities",
            "type": "array"
          },
          "quantity_units": {
            "items": {
              "$ref": "#/components/schemas/EnumOption"
            },
            "title": "Quantity Units",
            "type": "array"
          },
          "severities": {
            "items": {
              "$ref": "#/components/schemas/EnumOption"
            },
            "title": "Severities",
            "type": "array"
          },
          "user_roles": {
            "items": {
              "$ref": "#/components/schemas/EnumOption"
            },
            "title": "User Roles",
            "type": "array"
          }
        },
        "required": [
          "user_roles",
          "complaint_statuses",
          "severities",
          "priorities",
          "complaint_sources",
          "complaint_types",
          "dosage_forms",
          "quantity_units",
          "document_kinds"
        ],
        "title": "VocabularyResponse",
        "type": "object"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "description": "Pharmaceutical customer complaint management with AI-assisted intake, triage and investigation support.",
    "title": "AI-Powered Customer Complaint Management System",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "description": "Standard OAuth2 password flow. `username` carries the email address.\n\nThis is the endpoint Swagger's Authorize button uses.",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_login_api_v1_auth_login_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Token"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Log in (OAuth2 form)",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/login/json": {
      "post": {
        "description": "JSON login, used by the React client.",
        "operationId": "login_json_api_v1_auth_login_json_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Token"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Log in (JSON)",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "description": "The caller, with the permissions their role holds.\n\nThe frontend uses the permission list to hide controls. It is a usability\naffordance, not a security boundary - every endpoint checks independently.",
        "operationId": "me_api_v1_auth_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserWithPermissions"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Current user",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/users": {
      "get": {
        "description": "Used to populate investigator-assignment dropdowns.",
        "operationId": "list_users_api_v1_auth_users_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UserRead"
                  },
                  "title": "Response List Users Api V1 Auth Users Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "List users",
        "tags": [
          "auth"
        ]
      },
      "post": {
        "description": "Admin-only. Emails are normalised to lowercase so login is case-insensitive.",
        "operationId": "create_user_api_v1_auth_users_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Create user",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/meta/vocabulary": {
      "get": {
        "description": "Single source of dropdown options for the entire frontend.",
        "operationId": "get_vocabulary_api_v1_meta_vocabulary_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VocabularyResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "All domain enums",
        "tags": [
          "meta"
        ]
      }
    },
    "/health": {
      "get": {
        "description": "Unauthenticated liveness check used by Docker, CI and the smoke tests.",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Health Health Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Liveness probe",
        "tags": [
          "system"
        ]
      }
    }
  }
}